实体统一集成basedomain,统一新增is_del字段软删除,其他等修复

+++ yshop-shop/src/main/java/co/yixiang/modules/activity/service/YxStoreCombinationService.java
@@ -46,5 +46,5 @@ public interface YxStoreCombinationService extends BaseService<YxStoreCombinatio
      */
     void download(List<YxStoreCombinationDto> all, HttpServletResponse response) throws IOException;
 
-    void onSale(Integer id, int status);
+    void onSale(Long id, int status);
 }
